remove popup without cursor control
