+++ src/ui/components/mod.rs
@@ -1,7 +1,6 @@
 pub mod input;
 pub mod input_popup;
 pub mod layout;
-pub mod popup;
 pub mod popup_host;
 pub mod popup_with_cursor;
 pub mod view_switcher;
